requirements.txt
streamlit
torch
torchaudio
faster-whisper
openai
streamlit-audiorecorder
pydub
